+++ docs/changes.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+Version 3.0.0 (unreleased)
+========================
+ - Remove support for Python 2.6.
+ - Add support for Graph API version 2.8.
+ - Remove support for Graph API version 2.1.
+ - Change default Graph API version to 2.2.
+ - Add support for requests' sessions.
+
 Version 2.0.0 (2016-08-08)
 ==============================
  - Add support for Graph API versions 2.6 and 2.7.